# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da API do W-AZAP",
    "version": "1.6.4",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor da API"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave de API gerada no painel (Webhooks e API)"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "Cookie de login do painel (somente navegador)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error occurred"
          },
          "error": {
            "type": "string",
            "example": "Detailed error info"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autenticado: chave de API ausente ou inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Requisição inválida: parâmetros incorretos",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Invalid request parameters",
              "error": "Invalid request parameters"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "paths": {
    "/media/{filename}": {
      "get": {
        "tags": [
          "Mídia"
        ],
        "summary": "Baixar arquivo de mídia",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conteúdo do arquivo"
          }
        }
      }
    },
...
    "/messages/{sessionId}/download/{messageId}/media": {
      "get": {
        "tags": [
          "Mensagens"
        ],
        "summary": "Baixar a mídia de uma mensagem",
        "description": "Baixa a mídia de uma mensagem (arquivo binário direto ou redirecionamento)",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo de mídia binário",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Proibido: sem acesso à sessão ou à mensagem"
          },
          "404": {
            "description": "Mensagem não encontrada ou sem mídia"
          }
        }
      }
    },
...
    "/media": {
      "get": {
        "tags": [
          "Mídia"
        ],
        "summary": "Listar arquivos de mídia",
        "description": "Lista os arquivos de mídia baixados das sessões do usuário, com dados do remetente. O SUPERADMIN vê os arquivos de todas as sessões.",
        "responses": {
          "200": {
            "description": "Arquivos de mídia",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Media files fetched successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "files": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "vendas-01-3EB0B78A.jpg"
                              },
                              "size": {
                                "type": "integer",
                                "example": 52341
                              },
                              "type": {
                                "type": "string",
                                "example": "image"
                              },
                              "sessionId": {
                                "type": "string",
                                "example": "vendas-01"
                              },
                              "from": {
                                "type": "string",
                                "example": "5511987654321@s.whatsapp.net"
                              },
                              "fromMe": {
                                "type": "boolean",
                                "example": false
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "url": {
                                "type": "string",
                                "example": "/api/media/vendas-01-3EB0B78A.jpg"
                              }
                            }
                          }
                        },
                        "totalSize": {
                          "type": "integer",
                          "example": 52341
                        },
                        "totalCount": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Mídia"
        ],
        "summary": "Excluir arquivos de mídia",
        "description": "Exclui vários arquivos de uma vez. Cada arquivo é conferido: só é possível excluir mídia das sessões às quais você tem acesso. Arquivos sem sessão identificável exigem SUPERADMIN.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "filenames"
                ],
                "properties": {
                  "filenames": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "vendas-01-3EB0B78A.jpg"
                    ],
                    "description": "Nomes dos arquivos (sem caminho)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado da exclusão (`deleted`, `failed`, `errors`)"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
...
}